+++ apps/member-profile/app/routes/_profile.resume-books.$id.tsx
@@ -29,13 +29,13 @@ import {
   getResumeBookSubmission,
   listResumeBookSponsors,
   submitResume,
-} from '@oyster/core/resumes';
+} from '@oyster/core/resume-books';
 import {
   RESUME_BOOK_CODING_LANGUAGES,
   RESUME_BOOK_JOB_SEARCH_STATUSES,
   RESUME_BOOK_ROLES,
   SubmitResumeInput,
-} from '@oyster/core/resumes/types';
+} from '@oyster/core/resume-books/types';
 import { db } from '@oyster/db';
 import { FORMATTED_RACE, Race, WorkAuthorizationStatus } from '@oyster/types';
 import {
