+++ packages/core/package.json
@@ -43,9 +43,10 @@
     "./remix": "./src/modules/remix.ts",
     "./resources": "./src/modules/resource/index.ts",
     "./resources/server": "./src/modules/resource/index.server.ts",
-    "./resumes": "./src/modules/resume/resume.core.ts",
-    "./resumes/types": "./src/modules/resume/resume.types.ts",
-    "./resumes/ui": "./src/modules/resume/resume.ui.tsx",
+    "./resume-books": "./src/modules/resume-book/resume-book.ts",
+    "./resume-books/types": "./src/modules/resume-book/resume-book.types.ts",
+    "./resume-books/ui": "./src/modules/resume-book/resume-book.ui.tsx",
+    "./resume-review": "./src/modules/resume-review/resume-review.ts",
     "./scholarships": "./src/modules/scholarship/scholarship.core.ts",
     "./scholarships/ui": "./src/modules/scholarship/scholarship.ui.tsx",
     "./slack": "./src/modules/slack/slack.ts",
